+++ tests/produce_consume_base_test.rs
@@ -126,7 +126,10 @@ fn test_produce_consume_with_timestamp() {
     // Lookup the offsets
     let tpl = consumer.offsets_for_timestamp(999999, 10000).unwrap();
     let tp = tpl.find_partition(&topic_name, 0).unwrap();
+    assert_eq!(tp.topic(), topic_name);
     assert_eq!(tp.offset(), Offset::Offset(100));
+    assert_eq!(tp.partition(), 0);
+    assert_eq!(tp.error(), Ok(()));
 }
 
 #[test]